requirements.txt
folium==0.14.0
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
branca==0.6.0
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
python-dotenv==1.0.0
requests==2.31.0 